Use size_t for string length in base64

+++ src/base64.hxx
@@ -5,7 +5,7 @@
 #include <string>
 
 std::string base64_encode(unsigned char const *bytes_to_encode,
-						  unsigned int in_len);
+						  size_t in_len);
 std::string base64_decode(std::string const& base64_string);
 
 #endif
